requirements.txt
fastapi[standard]>=0.115.0,<0.116.0
pyjwt>=2.10.1,<2.11.0
passlib[bcrypt]>=1.7.4,<1.8.0
sqlmodel>=0.0.22,<0.1.0
asyncpg>=0.30.0,<0.31.0
greenlet>=3.1.1,<3.2.0
python-dotenv>=1.0.1,<1.1.0
uvicorn[standard]>=0.25.0,<0.26.0